# swagger.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "Small BLOG API ",
    "description": "API where people share ideas, small blog using PERN.",
    "contact": { "email": "bassel2053@gmail.com" },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "host": "localhost:3000",
  "basePath": "/",
...
  "schemes": ["http"],
  "produces": ["application/json"],
  "paths": {
    "/users/signup": {
      "post": {
        "tags": ["auth"],
        "summary": "Create a new user",
        "description": "creating new users in blog",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "operationId": "addUser",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Created user object",
            "required": true,
            "schema": { "$ref": "#/definitions/NewUser" }
          }
        ],
        "responses": {
          "201": {
            "description": "User created successfully",
            "schema": {
              "type": "object",
              "properties": {
                "message": { "type": "string", "example": "user created" },
                "user": {
                  "type": "object",
                  "properties": {
                    "email": {
                      "type": "string",
                      "example": "test@example.com"
                    },
                    "name": { "type": "string", "example": "john" },
                    "id": { "type": "number", "example": 10 }
                  }
                }
              }
            }
          },
          "422": { "description": "Validation error happened" }
        }
      }
    },
    "/users/login": {
      "post": {
        "tags": ["auth"],
        "summary": "User signin",
        "description": "User login to get access to blog resources like posts, comments and profiels NOTE: Must confirm E-mail to login.",
        "operationId": "login",
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Login object",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "test@example.com"
                },
                "password": { "type": "string", "example": "000000" }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User is authenticated",
            "schema": {
              "type": "object",
              "properties": {
                "name": { "type": "string", "example": "john" },
                "accessToken": {
                  "type": "string"
                }
              }
            }
          },
          "401": { "description": "E-mail is not confirmed" },
          "404": {
            "description": "There is no user with such a combination of email and password"
          },
          "422": { "description": "Validation error happened" }
        }
      }
    },
    "/users/forgot-password": {
      "post": {
        "tags": ["auth"],
        "summary": "Forgot password",
        "description": "Sending a request to get an e-mail which provides a code and a link, which can be used to be able to reset the password",
        "operationId": "forgot",
        "parameters": [
          {
            "in": "body",
            "name": "email",
            "description": "forgotten account email",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "email": {
                  "type": "string",
                  "example": "test@example.com"
                }
              }
            }
          }
        ],
        "responses": {
          "200": {
            "description": "User will recieve email to reset password",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "An E-mail has been sent."
                }
              }
            }
          },
          "404": {
            "description": "There is no user with such an email"
          },
          "422": { "description": "Validation error happened" }
        }
      }
    },
    "/users/check-reset": {
      "post": {
        "tags": ["auth"],
        "summary": "Check reset code",
        "description": "Checks if the user provided proper reset token and code to gain access to rese the password",
        "operationId": "checkCode",
        "parameters": [
          {
            "in": "body",
            "name": "code",
            "description": "reset code",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "code": {
                  "type": "string",
                  "example": "123456"
                }
              }
            }
          },
          {
            "in": "query",
            "name": "token",
            "description": "token for reset request",
            "required": true,
            "type": "string"
          },
          {
            "in": "query",
            "name": "id",
            "description": "user id",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "The user provided correct reset code.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "The reset code has been verified you are ready to change the password"
                }
              }
            }
          },
          "404": {
            "description": "There reset token expired token or wrong code was provided"
          },
          "422": { "description": "Validation error happened" }
        }
      }
    },
    "/users/reset-password": {
      "post": {
        "tags": ["auth"],
        "summary": "Provide new password",
        "description": "The user will provide new password after submitting reset request and providing correct reset token and code.",
        "operationId": "reset",
        "parameters": [
          {
            "in": "body",
            "name": "code",
            "description": "reset code",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "password": {
                  "type": "string",
                  "example": "new123"
                },
                "confirmPassword": {
                  "type": "string",
                  "example": "new123"
                }
              }
            }
          },
          {
            "in": "query",
            "name": "id",
            "description": "user id",
            "required": true,
            "type": "number"
          }
        ],
        "responses": {
          "200": {
            "description": "user sucessfully created a new password.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "You changed your password successfully."
                }
              }
            }
          },
          "403": {
            "description": "Password reset code is not verified and user won't be allowed to change the password."
          },
          "422": { "description": "Validation error happened" }
        }
      }
    },
    "/users/confirm/{token}": {
      "get": {
        "tags": ["auth"],
        "summary": "user email confirmation",
        "description": "The user will provide new password after submitting reset request and providing correct reset token and code.",
        "operationId": "confirmEmail",
        "parameters": [
          {
            "in": "path",
            "name": "token",
            "description": "confirmation token",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "302": {
            "description": "Redirecting user after confirming the email, to be able to login."
          },
          "403": {
            "description": "User is trying to confirm an email which is alread confirmed or a user that doesn't exists."
          }
        }
      }
    },
...
    "/token/refresh": {
      "get": {
        "tags": ["auth"],
        "summary": "refresh user access token",
        "description": "Generate a new access token for the user form refresh token to be able to preserve consistent login for users in the blog.",
        "operationId": "refreshToken",
        "responses": {
          "200": {
            "description": "a new access token is generated and is sent to the client.",
            "schema": {
              "type": "object",
              "properties": {
                "message": {
                  "type": "string",
                  "example": "user token refreshed"
                },
                "name": {
                  "type": "string",
                  "example": "john"
                },
                "accessToken": {
                  "type": "string"
                }
              }
            }
          },
          "401": {
            "description": "User is not allowed to generate a new access token if there is no refresh token"
          },
          "403": {
            "description": "Refresh token reuse is detected and the all of the user refresh tokens will be deleted"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "NewUser": {
      "type": "object",
      "required": ["email", "password", "confirmPassword", "name"],
      "properties": {
        "email": {
          "type": "string",
          "description": "unique email for every user account",
          "example": "test@example.com"
        },
        "password": {
          "type": "string",
          "description": "the user password should be 6-16 characters",
          "example": "000000"
        },
        "confirmPassword": {
          "type": "string",
          "description": "it needs to match the password",
          "example": "000000"
        },
        "name": {
          "type": "string",
          "description": "name which acts as username, need to be unique",
          "example": "john"
        }
      }
    },
...
  }
}